MATLAB based Speaker Verification System
In this project first the system is trained and then it is used to
recognize the speaker from given set, hence its a Speaker verification
system and not a Speaker Identification system.
The file combined.m contains final code.
cepstrum.m contains the code for calculating Mel Frequency Cepstrum
Coefficients.
data_collection.m contains the code for collecting data and training the
system.